Move prop types from types.ts to individual components
