+++ package.json
@@ -41,9 +41,10 @@
     "@emotion/styled": "^11.11.5",
     "@mui/icons-material": "^5.16.0",
     "@mui/material": "^5.14.19",
-    "gray-matter": "^4.0.3",
     "chart.js": "^4.4.7",
+    "chartjs-adapter-moment": "^1.0.1",
     "colorjs.io": "^0.5.2",
+    "gray-matter": "^4.0.3",
     "react-chartjs-2": "^5.3.0",
     "react-markdown": "^9.0.1",
     "react-wrap-balancer": "^1.1.1",
