+++ src/Table/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Table";
+export { default, type TableProps } from "./Table";
