+++ src/index.ts
@@ -2,8 +2,9 @@ export { default as ArticleCard } from "./ArticleCard";
 export { default as HorizontalArticleCard } from "./HorizontalArticleCard";
 export { default as TopStyledBlock } from "./TopStyledBlock";
 export { default as Article } from "./Article";
-export { default as PieChart } from "./PieChart";
-export { default as Table } from "./Table";
+export { default as PieChart, type PieChartData } from "./PieChart";
+export { default as Table, type TableProps } from "./Table";
 export { default as ColorBar } from "./ColorBar";
+export { default as BigNumber, type BigNumberProps } from "./BigNumber";
 export { default as TimeBar } from "./TimeBar";
 export * from './utils';
